Update MainWindowController with new app config sections
